[DC-983] Move log_traceback to app_errors
